Added events page skeleton, refactored team page, updated title, added data folder.
