+++ src/data/events.js
@@ -1,0 +1,37 @@
+export const events = [
+  {
+    id: 1,
+    title: "Event 1",
+    date: "January 2024",
+    description: "Description of Event 1",
+    images: [
+      {
+        url: "/pics_png/GALLERY1.jpg",
+        description: "Description for image 1"
+      },
+      {
+        url: "/pics_png/GALLERY2.jpg",
+        description: "Description for image 2"
+      },
+      // Add more images as needed
+    ]
+  },
+  {
+    id: 2,
+    title: "Event 2",
+    date: "February 2024",
+    description: "Description of Event 2",
+    images: [
+      {
+        url: "/pics_png/GALLERY5.jpg",
+        description: "Description for image 1"
+      },
+      {
+        url: "/pics_png/GALLERY6.jpg",
+        description: "Description for image 2"
+      },
+      // Add more images as needed
+    ]
+  },
+  // Add more events as needed
+];
