+++ src/data/team.js
@@ -1,0 +1,252 @@
+export const teams = [
+  {
+    title: "Executive Board",
+    leads: [
+      {
+        name: "Aishwarya Alichella",
+        role: "President",
+        image: "team_pics/aishwarya.webp",
+      },
+      {
+        name: "Eshwar",
+        role: "Vice President",
+        image: "team_pics/eshwar.webp",
+      },
+    ],
+    members: [
+      {
+        name: "Sandhya",
+        role: "General Secretary",
+        image: "team_pics/",
+      },
+      {
+        name: "Ankitha",
+        role: "Joint Secretary",
+        image: "team_pics/",
+      },
+      {
+        name: "Kaif Shaik",
+        role: "Treasurer",
+        image: "team_pics/",
+      },
+      {
+        name: "Vigneshwar",
+        role: "Organising Lead",
+        image: "team_pics/",
+      },
+      {
+        name: "Eshika",
+        role: "Senior Executive Board",
+        image: "team_pics/",
+      },
+      {
+        name: "Bindu Priya",
+        role: "Senior Executive Board",
+        image: "team_pics/",
+      },
+      {
+        name: "Raaga Samanvitha",
+        role: "Ethics & Operations Head",
+        image: "team_pics/",
+      },
+      {
+        name: "Vaishnavi V",
+        role: "Sustainable Operations Head",
+        image: "team_pics/vaishnavi_v.webp",
+      },
+      {
+        name: "Jerusha",
+        role: "Alumni Relations",
+        image: "team_pics/",
+      },
+    ],
+  },
+  {
+    title: "Research And Development Team",
+    leads: [
+      {
+        name: "Srujan M",
+        role: "Research And Development Team Head",
+        image: "team_pics/",
+      },
+    ],
+    members: [
+      {
+        name: "Pranav",
+        image: "",
+      },
+      {
+        name: "Police Sreeja",
+        image: "team_pics/",
+      },
+      {
+        name: "K Rekha",
+        image: "team_pics/rekha.webp",
+      },
+      {
+        name: "Anji Reddy B",
+        image: "team_pics/anji.webp",
+      },
+    ],
+  },
+  {
+    title: "Technical Team",
+    leads: [
+      {
+        name: "Amrutha Sri",
+        role: "Technical Team Head",
+        image: "team_pics/amrutha.webp",
+      },
+    ],
+    members: [
+      {
+        name: "Pavan Kumar",
+        image: "team_pics/pavan.webp",
+      },
+      {
+        name: "Deekshitha Reddy",
+        image: "team_pics/deekshitha.webp",
+      },
+    ],
+  },
+  {
+    title: "Design Team",
+    leads: [
+      {
+        name: "Poornima Guda",
+        role: "Design Team Head",
+        image: "team_pics/poornima.webp",
+      },
+    ],
+    members: [
+      {
+        name: "Kavya Sri",
+        image: "https://via.placeholder.com/200",
+      },
+      {
+        name: "Pranathi",
+        image: "team_pics/pranathi.webp",
+      },
+      {
+        name: "Uma Sri",
+        image: "https://via.placeholder.com/200",
+      },
+      {
+        name: "Tella Vaishnavi",
+        image: "team_pics/vaishnavi.webp",
+      },
+      {
+        name: "Vamshi Krishna",
+        image: "team_pics/vamshi_design.webp",
+      },
+    ],
+  },
+  {
+    title: "Marketing Team",
+    leads: [
+      {
+        name: "Akshitha T",
+        role: "Marketing Team Head",
+        image: "team_pics/akshitha.webp",
+      },
+    ],
+    members: [
+      {
+        name: "Abhiram",
+        image: "team_pics/abhiram.webp",
+      },
+      {
+        name: "Sahithi",
+        image: "team_pics/sahithi.webp",
+      },
+    ],
+  },
+  {
+    title: "Content Team",
+    leads: [
+      {
+        name: "Sonal G",
+        role: "Content Team Head",
+        image: "https://via.placeholder.com/200",
+      },
+    ],
+    members: [
+      {
+        name: "Thanughna D",
+        image: "team_pics/thanughana.webp",
+      },
+      {
+        name: "Sukruth B",
+        image: "team_pics/sukruth.webp",
+      },
+      {
+        name: "S. F. Cecilia",
+        image: "team_pics/",
+      },
+      {
+        name: "Keerthana",
+        image: "team_pics/",
+      },
+    ],
+  },
+  {
+    title: "Public Relations Team",
+    leads: [
+      {
+        name: "Shyamala",
+        role: "Public Relations Team Head",
+        image: "https://via.placeholder.com/200",
+      },
+    ],
+    members: [
+      {
+        name: "Sumad",
+        image: "team_pics/sumad.webp",
+      },
+      {
+        name: "Jahnavi",
+        image: "https://via.placeholder.com/200",
+      },
+      {
+        name: "Agrathi Sheela",
+        image: "team_pics/agrathi.webp",
+      },
+      {
+        name: "Sri Charan Raj",
+        image: "https://via.placeholder.com/200",
+      },
+    ],
+  },
+  {
+    title: "Social Media Team",
+    leads: [
+      {
+        name: "Abhay",
+        role: "Social Media Team Head",
+        image: "team_pics/abhay.webp",
+      },
+    ],
+    members: [
+      {
+        name: "Ayesha Shaik",
+        image: "team_pics/",
+      },
+      {
+        name: "Vyshali Katta",
+        image: "team_pics/",
+      },
+      {
+        name: "Shiva Sai",
+        image: "team_pics/",
+      },
+      {
+        name: "Ram Swaroop",
+        image: "team_pics/ram",
+      },
+      {
+        name: "Siri Chandana Reddy",
+        image: "team_pics/siriChandana.webp",
+      },
+    ],
+  },
+];
